pyproject.toml
[tool.black]
line-length = 79
target-version = ["py39"]


[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
testpaths = [
    "tests",
]

# =============================================================================
# PACKAGING
# =============================================================================

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "gvt_scripts"
version = "0.2"
authors = [{name = "Juan B Cabral & GVT-CONAE", email = "jbcabral@unc.edu.ar"}]
readme = "README.md"
license = {file = "LICENSE.txt"}
description = "Multiple scripts for the GVT"
keywords = [
    "GVT",
    "CONAE",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering",
]
urls = { Homepage = "https://github.com/leliel12/gvt_scripts", Repository = "https://github.com/leliel12/gvt_scripts" }
dependencies = [
    # GENERAL
    "attrs",
    "typer",
    "importlib_metadata",
    "pyyaml",
    "orjson",
    "dicttoxml",
    "tomli_w",
    "joblib",
    "peewee",
    "python-dateutil",

    # shp_metadata
    "pyproj",
    "dbfread",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["gvt_scripts", "gvt_scripts.*"]
namespaces = false

[project.scripts]
search-shp-dir = "gvt_scripts.search_shp_dir:main"